First
This is the base code that was tested at the end of the year with Simon. We have yet to implement any sort of LEDs, buttons, or switches. While we have all the other base functions we need, we still have to work on the code for the competition specifically on the stopping mech part.